move into subfolders specifically for factories
this is to prep for creating a new folder for utils, in which a localstorage module will contain helper functions we are about to write for this branch
